add mock record exporter
