+++ ptolemy-py/python/ptolemy_client/__init__.py
@@ -2,7 +2,8 @@
 
 # pylint: disable=no-name-in-module
 
-from .client import connect, Ptolemy
+from .client.client import Ptolemy, connect
+from .trace import Trace
 
 # Used by python-semantic-release
 __version__ = "0.0.0-test.4+7da95d5"
